divided concepts of comms and controller
now a comms interface is a simple class which returns a more complex
controller upon openController(). This allows the controller
implementation to have hard dependencies on information queried from the
actual device.

it gives REST interface over the controller a degree of separation from
the comms API, so as to ease facilitation of other types of device by
factoring it out when necessary.