tons of work on persisted state/spectra/experiments...
Spectra now have their processing configured inline rather than being
represented part of the experiment tree, and the persisted state system
has been enhanced to support this.

Begun to factor out more general and useful experiment implementation
based on persisted state from the XML implementation.

Begun to think about "snippets" of experiments in memory without
persistence and how they might help with editors acting on a view of the
experiment not directly on persisted state ... and in turn how this
might relate to a templating system for users to paste in common bits of
experiment trees.