Lots of work on the experimemnt model
- Refactoring of experiment path API to reify distinction between
absolute and relative.

- Refactoring of State API to separate concept of an accessor (a
converter between a unit of state and a java type) and a MapIndex (an
accessor and a map key to extract/modify state at that id according to
the accessor), also added ListIndex

- Added event types for experiment modifications

- Refactored scheduler and implemented Conflicts API to report
information on whether a scheduled procedure conflicts with existing
gathered results

- Overhaul of Conductor API. Instead of being parametric with a single
"variables" type which contains all information, we can attach any
number of Variable objects to an experiment to separate units of
configuration out. Also removed the system of mutable variables which
automatically propagate changes into the underlying state, as this was
not neatly compatible with immutable underlying instructions and atomic
updates to experiment steps in the mutable view. This means there is no
more ConfigurationContext, which is simpler for implementors of
conductors.

- Refined concept of indirect dependencies on an instruction and how
they are derived from the instruction state and the containing procedure

- "Add Step" actions in UI are no longer automatically generated from
the available conductor service objects. Instead we have a dedicated
ExperimentStepProvider in the UI layer which can provide extra UI
interaction and localized names without polluting the business layer

- Bubbled most of the changes up into the UI layer so we can actually
have a working product again...