improvements to observables, devices, and stages
- observables and observable values now take suppliers for failure
exceptions so that the stack trace can more accurately reflect the
location of failure
- taken toValue out of observables, semantics were awkward
- removed get and tryGet from observables, now just use getNext
- connection state now renamed device state and more richly reflects the
status of a device, including whether or not it is available to be
acquired
- immediate acquireControl now returns an optional instead of throwing
when unavailable.
- DeviceImpl now factors out much more of the boilerplate of creating a
device, including setting accessibility of underlying system and locking
on the controller
- new AbstractingDevice created for devices which consume and abstract
over one or more other devices, requiring control of them.
- StageAxis changed to be a device, meaning it has a richer API for
discovering and reacting to changes in availability and locking control
- ComposedStage reimplemented based on AbstractingDevice
- introduced sample plates and sample areas for saint stage device
- saint stage executor hierarchical, node containing plate settings has
child nodes containing area settings
- saint stage is separated into two devices, the plate device which has
control over sample areas, and the area device which has control over
offset within an area.
- new arithmatic operators in XYCoordinate