Included "shared method" feature for experiment designs, for more convenient reuse of common experiment elements.
Renamed ExperimentDefinition to ExperimentDesign, to better reflect the
purpose of the API within the wider system.

Made Procedures/Instructions validate on construction and report their
dependencies. This means errors can be caught earlier and simplifies a
lot of the conductor/scheduler logic downstream.

Added some javadocs to Procedure/ExperimentDesign/Experiment etc. to
elucidate on the motivations behind some of the important API design
choices.